django/requirements.txt
backports-abc>=0.5
chardet>=2.3.0
click>=6.7
decorator>=4.0.10
Django>=2.0.4
django-anymail>=0.11.1
django-appconf>=1.0.2
django-compat>=1.0.15
django-debug-toolbar>=1.6
django-extensions>=2.0.6
django-hijack>=2.1.4
django-imagekit>=3.3
django-import-export>=1.0.1
django-ratelimit>=1.0.0
django-rq>=0.9.5
django-widget-tweaks>=1.4.1
djangorestframework>=3.8.2
entrypoints>=0.2.2
httpie>=0.9.6
IPy>=0.83
ipykernel>=4.5.0
ipython>=5.1.0
ipython-genutils>=0.1.0
ipywidgets>=5.2.2
Jinja2>=2.8
jsonschema>=2.5.1
jupyter-client>=4.4.0
jupyter-core>=4.2.0
lxml>=3.6.4
MarkupSafe>=0.23
mistune>=0.7.3
mpmath>=0.19
nbconvert>=4.2.0
nbformat>=4.1.0
notebook>=4.2.3
numpy>=1.11.2
pexpect>=4.2.1
pickleshare>=0.7.4
pilkit>=1.1.13
Pillow>=3.3.1
pprofile>=1.9.2
prompt-toolkit>=1.0.8
ptyprocess>=0.5.1
pydotplus>=2.0.2
Pygments>=2.1.3
pygraphviz>=1.3.1
pyparsing>=2.1.9
PyPDF2>=1.26.0
pytz>=2016.7
pyzmq>=16.0.0
redis>=2.10.5
requests>=2.11.0
rq>=0.8.0
selenium>=2.53.6
simplegeneric>=0.8.1
six>=1.11.0
sqlparse>=0.2.1
sympy>=1.0
terminado>=0.6
tornado>=4.4.2
traitlets>=4.3.1
wcwidth>=0.1.7
widgetsnbextension>=1.2.6
XlsxWriter>=0.9.4
xmljson>=0.1.6
xmltodict>=0.11.0
